% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{cbind_list}
\alias{cbind_list}
\title{Bind together lists, not losing track of filenames}
\usage{
cbind_list(data_list, cbind_vals, cbind_name)
}
\description{
Bind together lists, not losing track of filenames
}
