% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{experiment_boxplots}
\alias{experiment_boxplots}
\title{Boxplots across experiment configurations}
\usage{
experiment_boxplots(mcombined)
}
\arguments{
\item{mcombined}{[data.frame] The melted data frame including true and
sampled values across experimental configurations.}
}
\value{
p [ggplot] The ggplot object containing the sampled values across
  configurations.
}
\description{
Boxplots across experiment configurations
}
