% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{experiment_contours}
\alias{experiment_contours}
\title{Contour plots across topic components}
\usage{
experiment_contours(combined)
}
\arguments{
\item{combined}{[data.frmae] The unmelted data.frame including true and
sampled values across experimental configurations}

\item{return}{p [ggplot] The ggplot object containing samples across
configurations.}
}
\description{
Contour plots across topic components
}
