% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{experiment_pi}
\alias{experiment_pi}
\title{Permutation to align topics across simulations}
\usage{
experiment_pi(index, dimension, estimate, truth)
}
\description{
Permutation to align topics across simulations
}
