% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_model}
\alias{fit_model}
\title{Fit an LDA model using STAN}
\usage{
fit_model(stan_data, stan_file, keep_samples = FALSE, use_vb = TRUE, ...)
}
\arguments{
\item{stan_data}{[list] A list giving required values for the data field in
the stan file.}

\item{stan_file}{[string] The path to the STAN file describing the
LDA model.}

\item{keep_samples}{[bool] Whether to keep samples from the model
posterior. Otherwise only returns (sampled) posterior mean.}

\item{use_vb}{[bool] Should we use variational bayes? If false,
uses gibbs sampling.}
}
\value{
result [list of data.table] A list of data.tables giving
  means for beta, theta, and (optionally) samples for beta and
  theta.
}
\description{
This is a wrapper for rstan::vb() and rstan::extract() which runs
and extracts summary information from an LDA model.
}

