% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{generate_params}
\alias{generate_params}
\title{Simulate parameters for LDA}
\usage{
generate_params(D, K, V, alpha0, gamma0)
}
\arguments{
\item{D}{[int] The number of documents.}

\item{K}{[int] The number of topics.}

\item{V}{[int] The number of terms.}

\item{alpha0}{[numeric vector] The location / concentration for the mixture
variables theta_{i}.}

\item{gamma0}{[numeric vector] The location / concentration for the topic
variables beta_{k}.}
}
\value{
x [list] A list with components $theta and $beta containing the
  simulated mixture components and topic matrices, respectively. For theta,
  the rows are individual documents, for beta, the rows are terms. Both
  matrices have K columns.
}
\description{
Simulate parameters for LDA
}
