% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{get_bootstraps}
\alias{get_bootstraps}
\title{Combine bootstrap samples for a parameter}
\usage{
get_bootstraps(metadata, param, variable = "v")
}
\arguments{
\item{metadata}{[data.frame] The metadata data frame, including all paths to
data generated by the simulation.}

\item{param}{[string] The name of the parameter to draw posterior samples for}
}
\value{
bootstraps [data.frame]
}
\description{
This retrieves all posterior samples for a parameter, across many
experimental configurations.
}
