% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{get_samples}
\alias{get_samples}
\title{Combine posterior samples for a parameter}
\usage{
get_samples(metadata, param, var_names = NULL)
}
\arguments{
\item{metadata}{[data.frame] The metadata data frame, including all paths to
data generated by the simulation.}

\item{param}{[string] The name of the parameter to draw posterior samples for}
}
\value{
samples [data.frame]
}
\description{
This retrieves all posterior samples for a parameter, across many
experimental configurations.
}

