% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{get_truth_data}
\alias{get_truth_data}
\title{Read true parameter values from file}
\usage{
get_truth_data(metadata, param, variable = "v")
}
\arguments{
\item{metadata}{[data.frame] The metadata data frame, including all paths to
data generated by the simulation.}

\item{param}{[string] The name of the parameter to draw posterior samples for}
}
\value{
truth [data.frame]
}
\description{
Read true parameter values from file
}

