% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Simulate from a Dirichlet Distribution}
\usage{
rdirichlet(n, alpha)
}
\arguments{
\item{n}{[int] The number of dirichlet samples to draw}

\item{alpha}{[numeric vector] The alpha location / concentration parameter
for the Dirichlet.}
}
\value{
x [matrix] A matrix whose rows are samples from a Dirichlet(alpha).
}
\description{
Simulate from a Dirichlet Distribution
}
